refactor(recording): split module internals
Split the large recording module into smaller internal modules:

- src/recording/types.rs: All public types (RecordingError, RecordingMode,
  RecordingBucket, ActiveRecording, RecordingFileEntry, RecordingsOverview,
  RecordingProcessingConfig) and internal types (ActiveProcess, ChapterEvent,
  ChannelMetadataCache)

- src/recording/files.rs: File/path helpers (sanitize_filename, build_recording_filename,
  format_filename_timestamp, validate_recording_filename, build_completed_recording_path,
  move_file_if_exists, find_file_by_name_recursive, list_recording_files,
  is_visible_recording_file, pin_marker_path_for_recording, is_recording_pinned,
  prune_completed_channel_dir)

- src/recording/nfo.rs: NFO generation helpers (write_episode_nfo_file,
  write_tvshow_nfo_file, next_same_day_suffix_index, parse_display_episode_suffix,
  xml_tag_value, xml_escape, datetime_from_unix, select_show_tags,
  read_channel_metadata_cache, write_channel_metadata_cache, update_channel_poster)

RecordingService intentionally remains in the root module to avoid broad
architecture changes. Playback asset generation also remains in the root module.

Public imports from crate::recording are preserved via re-exports.
All existing tests pass. No behavior changes intended.
