+++ src/recording.rs
@@ -4,12 +4,8 @@ use std::{
     path::{Path, PathBuf},
     process::Stdio,
     sync::Arc,
-    time::{SystemTime, UNIX_EPOCH},
 };
 
-use serde::{Deserialize, Serialize};
-use thiserror::Error;
-use time::{OffsetDateTime, format_description};
 use tokio::{process::Command, sync::RwLock};
 
 use crate::{
@@ -18,114 +14,18 @@ use crate::{
     twitch_auth::{HelixChannelMetadata, TwitchAuthService},
 };
 
-/// Typed error for recording operations.
-#[derive(Debug, Clone, PartialEq, Eq, Error)]
-pub enum RecordingError {
-    #[error("channel login cannot be empty")]
-    EmptyChannelLogin,
-    #[error("invalid quality")]
-    InvalidQuality,
-    #[error("recording already active")]
-    AlreadyActive,
-    #[error("recording not active")]
-    NotActive,
-    #[error("recording file not found")]
-    FileNotFound,
-    #[error("filename cannot be empty")]
-    EmptyFilename,
-    #[error("invalid filename")]
-    InvalidFilename,
-    #[error("recording delete failed: {0}")]
-    DeleteFailed(String),
-    #[error("recording pin failed: {0}")]
-    PinFailed(String),
-    #[error("recording unpin failed: {0}")]
-    UnpinFailed(String),
-    #[error("streamlink spawn failed: {0}")]
-    SpawnFailed(String),
-    #[error("recordings directory not writable: {0}")]
-    DirectoryNotWritable(String),
-    #[error("io error: {0}")]
-    Io(String),
-}
+mod files;
+mod nfo;
+mod types;
 
-const QUALITY_OPTIONS: [&str; 9] = [
-    "best", "source", "1080p60", "1080p", "720p60", "720p", "480p", "360p", "160p",
-];
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-struct ChannelMetadataCache {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    poster_url: Option<String>,
-    #[serde(default)]
-    tags: Vec<String>,
-}
+pub use types::{
+    ActiveRecording, RecordingBucket, RecordingError, RecordingFileEntry, RecordingMode,
+    RecordingProcessingConfig, RecordingsOverview,
+};
 
-#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum RecordingMode {
-    Manual,
-    Auto,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ActiveRecording {
-    pub channel_login: String,
-    pub quality: String,
-    pub started_at_unix: u64,
-    pub output_path: String,
-    pub pid: Option<u32>,
-    pub mode: RecordingMode,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RecordingFileEntry {
-    pub channel_login: String,
-    pub filename: String,
-    pub path_display: String,
-    pub status: String,
-    pub pinned: bool,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum RecordingBucket {
-    Completed,
-    Incomplete,
-}
-
-impl RecordingBucket {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Completed => "completed",
-            Self::Incomplete => "incomplete",
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RecordingsOverview {
-    pub active: Vec<ActiveRecording>,
-    pub completed: Vec<RecordingFileEntry>,
-    pub incomplete: Vec<RecordingFileEntry>,
-}
-
-#[derive(Debug)]
-struct ActiveProcess {
-    metadata: ActiveRecording,
-    stream_title: Option<String>,
-    last_observed_game: Option<String>,
-    pending_game: Option<String>,
-    pending_game_confirmations: u64,
-    chapter_events: Vec<ChapterEvent>,
-    child: tokio::process::Child,
-}
-
-#[derive(Debug, Clone)]
-struct ChapterEvent {
-    offset_secs: u64,
-    title: String,
-}
+use files::*;
+use nfo::*;
+use types::*;
 
 #[derive(Debug, Clone)]
 pub struct RecordingService {
@@ -138,13 +38,6 @@ pub struct RecordingService {
     chapter_min_gap_secs: u64,
     chapter_change_confirmations: u64,
     active: Arc<RwLock<HashMap<String, ActiveProcess>>>,
-}
-
-#[derive(Debug, Clone)]
-pub struct RecordingProcessingConfig {
-    pub ffmpeg_path: String,
-    pub chapter_min_gap_secs: u64,
-    pub chapter_change_confirmations: u64,
 }
 
 impl RecordingService {
@@ -884,467 +777,6 @@ impl RecordingService {
     }
 }
 
-fn write_episode_nfo_file(
-    channel_login: &str,
-    recording_path: &Path,
-    metadata: &ActiveRecording,
-    stream_title: Option<&str>,
-    genres: &[String],
-) -> Result<(), String> {
-    let Some(stem) = recording_path.file_stem().and_then(|value| value.to_str()) else {
-        return Err("failed to derive recording basename".to_string());
-    };
-
-    let started = datetime_from_unix(metadata.started_at_unix);
-    let season = started.year();
-    let month = started.month() as u8;
-    let day = started.day();
-    let base_episode: u16 = u16::from(month) * 100 + u16::from(day);
-    let aired = format!("{season:04}-{month:02}-{day:02}");
-
-    let channel_dir = recording_path
-        .parent()
-        .ok_or_else(|| "recording file has no parent directory".to_string())?;
-    let suffix_index = next_same_day_suffix_index(channel_dir, &aired, base_episode);
-    let episode_number = base_episode.saturating_add(suffix_index);
-    let display_episode = if suffix_index == 0 {
-        base_episode.to_string()
-    } else {
-        format!("{base_episode}-{suffix_index}")
-    };
-
-    let chosen_title = stream_title
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-        .unwrap_or_else(|| format!("{channel_login} stream {aired}"));
-    let title = if suffix_index == 0 {
-        chosen_title.clone()
-    } else {
-        format!("{chosen_title} ({display_episode})")
-    };
-    let mode = match metadata.mode {
-        RecordingMode::Manual => "manual",
-        RecordingMode::Auto => "auto",
-    };
-    let plot = format!(
-        "Twitch recording for {channel_login}. Title: {chosen_title}. Quality: {}. Mode: {mode}.",
-        metadata.quality
-    );
-    let uniqueid = format!(
-        "{}-{}",
-        sanitize_filename(channel_login),
-        metadata.started_at_unix
-    );
-    let genre_xml = genres
-        .iter()
-        .map(|genre| format!("  <genre>{}</genre>\n", xml_escape(genre)))
-        .collect::<String>();
-
-    let xml = format!(
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<episodedetails>\n  <title>{}</title>\n  <showtitle>{}</showtitle>\n  <season>{season}</season>\n  <episode>{episode_number}</episode>\n  <displayepisode>{}</displayepisode>\n  <aired>{aired}</aired>\n{}  <plot>{}</plot>\n  <uniqueid type=\"twitch\" default=\"true\">{}</uniqueid>\n</episodedetails>\n",
-        xml_escape(&title),
-        xml_escape(channel_login),
-        xml_escape(&display_episode),
-        genre_xml,
-        xml_escape(&plot),
-        xml_escape(&uniqueid)
-    );
-
-    let nfo_path = recording_path.with_file_name(format!("{stem}.nfo"));
-    fs::write(&nfo_path, xml)
-        .map_err(|error| format!("failed to write nfo file {}: {error}", nfo_path.display()))
-}
-
-fn next_same_day_suffix_index(channel_dir: &Path, aired: &str, episode: u16) -> u16 {
-    let Ok(entries) = fs::read_dir(channel_dir) else {
-        return 0;
-    };
-
-    let mut max_suffix: i32 = -1;
-    for entry in entries.flatten() {
-        let path = entry.path();
-        let is_nfo = path
-            .extension()
-            .and_then(|ext| ext.to_str())
-            .map(|ext| ext.eq_ignore_ascii_case("nfo"))
-            .unwrap_or(false);
-        if !is_nfo {
-            continue;
-        }
-
-        let Ok(content) = fs::read_to_string(&path) else {
-            continue;
-        };
-        if xml_tag_value(&content, "aired").as_deref() != Some(aired) {
-            continue;
-        }
-        let Some(episode_value) = xml_tag_value(&content, "episode") else {
-            continue;
-        };
-        if episode_value.trim().parse::<u16>().ok() != Some(episode) {
-            continue;
-        }
-
-        let display =
-            xml_tag_value(&content, "displayepisode").unwrap_or_else(|| episode_value.clone());
-        let parsed = parse_display_episode_suffix(&display, episode);
-        if parsed > max_suffix {
-            max_suffix = parsed;
-        }
-    }
-
-    (max_suffix + 1).max(0) as u16
-}
-
-fn parse_display_episode_suffix(display_episode: &str, episode: u16) -> i32 {
-    let trimmed = display_episode.trim();
-    let base = episode.to_string();
-    if trimmed == base {
-        return 0;
-    }
-
-    let Some(suffix) = trimmed.strip_prefix(&format!("{base}-")) else {
-        return 0;
-    };
-    suffix.parse::<i32>().unwrap_or(0)
-}
-
-fn xml_tag_value(content: &str, tag: &str) -> Option<String> {
-    let open = format!("<{tag}>");
-    let close = format!("</{tag}>");
-    let start = content.find(&open)? + open.len();
-    let end_rel = content[start..].find(&close)?;
-    Some(content[start..start + end_rel].trim().to_string())
-}
-
-fn datetime_from_unix(unix_secs: u64) -> OffsetDateTime {
-    i64::try_from(unix_secs)
-        .ok()
-        .and_then(|unix| OffsetDateTime::from_unix_timestamp(unix).ok())
-        .unwrap_or(OffsetDateTime::UNIX_EPOCH)
-}
-
-fn xml_escape(value: &str) -> String {
-    value
-        .replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-        .replace('\'', "&apos;")
-}
-
-fn build_completed_recording_path(
-    channel_dir: &Path,
-    channel_login: &str,
-    metadata: &ActiveRecording,
-    stream_title: Option<&str>,
-) -> PathBuf {
-    let started = datetime_from_unix(metadata.started_at_unix);
-    let season = started.year();
-    let month = started.month() as u8;
-    let day = started.day();
-    let base_episode: u16 = u16::from(month) * 100 + u16::from(day);
-    let aired = format!("{season:04}-{month:02}-{day:02}");
-    let season_dir = channel_dir.join(format!("Season {season}"));
-    let suffix = next_same_day_suffix_index(&season_dir, &aired, base_episode);
-    let episode_number = base_episode.saturating_add(suffix);
-
-    let title_slug = stream_title
-        .map(sanitize_filename)
-        .filter(|v| !v.is_empty())
-        .unwrap_or_else(|| "stream".to_string());
-    season_dir.join(format!(
-        "{}_S{season:04}E{episode_number:04}_{title_slug}.ts",
-        sanitize_filename(channel_login)
-    ))
-}
-
-fn write_tvshow_nfo_file(
-    channel_login: &str,
-    channel_dir: &Path,
-    metadata: &HelixChannelMetadata,
-    genres: &[String],
-) -> Result<(), String> {
-    let title = metadata.display_name.trim();
-    let plot = metadata
-        .description
-        .as_deref()
-        .filter(|v| !v.trim().is_empty())
-        .unwrap_or("Twitch channel recordings.");
-    let genre_xml = genres
-        .iter()
-        .map(|genre| format!("  <genre>{}</genre>\n", xml_escape(genre)))
-        .collect::<String>();
-    let xml = format!(
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<tvshow>\n  <title>{}</title>\n  <plot>{}</plot>\n  <status>Continuing</status>\n  <studio>{}</studio>\n{}  <thumb>poster.jpg</thumb>\n  <uniqueid type=\"twitch\" default=\"true\">twitch_{}</uniqueid>\n</tvshow>\n",
-        xml_escape(title),
-        xml_escape(plot),
-        xml_escape(title),
-        genre_xml,
-        xml_escape(channel_login)
-    );
-    let path = channel_dir.join("tvshow.nfo");
-    fs::write(&path, xml)
-        .map_err(|error| format!("failed to write tvshow.nfo {}: {error}", path.display()))
-}
-
-fn select_show_tags(
-    metadata: Option<&HelixChannelMetadata>,
-    cache: &ChannelMetadataCache,
-) -> Vec<String> {
-    let mut out: Vec<String> = Vec::new();
-    if let Some(meta) = metadata {
-        for tag in &meta.tags {
-            append_unique_tag(&mut out, tag);
-        }
-        if out.is_empty() {
-            for tag in &cache.tags {
-                append_unique_tag(&mut out, tag);
-            }
-        }
-        if out.is_empty()
-            && let Some(game) = meta.game.as_deref()
-        {
-            append_unique_tag(&mut out, game);
-        }
-    } else {
-        for tag in &cache.tags {
-            append_unique_tag(&mut out, tag);
-        }
-    }
-    out.truncate(10);
-    out
-}
-
-fn append_unique_tag(tags: &mut Vec<String>, raw: &str) {
-    let normalized = raw.trim();
-    if normalized.is_empty() {
-        return;
-    }
-    if tags.iter().any(|tag| tag.eq_ignore_ascii_case(normalized)) {
-        return;
-    }
-    tags.push(normalized.to_string());
-}
-
-fn read_channel_metadata_cache(channel_dir: &Path) -> ChannelMetadataCache {
-    let path = channel_dir.join(".metadata-cache.json");
-    let Ok(text) = fs::read_to_string(path) else {
-        return ChannelMetadataCache::default();
-    };
-    serde_json::from_str::<ChannelMetadataCache>(&text).unwrap_or_default()
-}
-
-fn write_channel_metadata_cache(
-    channel_dir: &Path,
-    cache: &ChannelMetadataCache,
-) -> Result<(), String> {
-    let path = channel_dir.join(".metadata-cache.json");
-    let payload = serde_json::to_string(cache)
-        .map_err(|error| format!("failed to encode channel metadata cache: {error}"))?;
-    fs::write(&path, payload).map_err(|error| {
-        format!(
-            "failed to write channel metadata cache {}: {error}",
-            path.display()
-        )
-    })
-}
-
-async fn update_channel_poster(
-    channel_dir: &Path,
-    http: &reqwest::Client,
-    metadata: &HelixChannelMetadata,
-    cache: &mut ChannelMetadataCache,
-) {
-    let Some(url) = metadata.profile_image_url.as_deref() else {
-        return;
-    };
-    if cache.poster_url.as_deref() == Some(url) {
-        return;
-    }
-    let Ok(response) = http.get(url).send().await else {
-        return;
-    };
-    if !response.status().is_success() {
-        return;
-    }
-    let Ok(bytes) = response.bytes().await else {
-        return;
-    };
-    let _ = fs::create_dir_all(channel_dir);
-    let poster_path = channel_dir.join("poster.jpg");
-    if fs::write(&poster_path, &bytes).is_ok() {
-        cache.poster_url = Some(url.to_string());
-    }
-}
-
-fn build_recording_filename(
-    channel: &str,
-    timestamp: u64,
-    quality: &str,
-    mode: RecordingMode,
-    stream_title: Option<&str>,
-) -> String {
-    let mode = match mode {
-        RecordingMode::Manual => "manual",
-        RecordingMode::Auto => "auto",
-    };
-    let safe_channel = sanitize_filename(channel);
-    let safe_quality = sanitize_filename(quality);
-    let formatted_timestamp = format_filename_timestamp(timestamp);
-    if let Some(title) = stream_title {
-        let safe_title = sanitize_filename(title);
-        if !safe_title.is_empty() {
-            return format!(
-                "{safe_channel}_{formatted_timestamp}_{safe_quality}_{mode}_{safe_title}.ts"
-            );
-        }
-    }
-    format!("{safe_channel}_{formatted_timestamp}_{safe_quality}_{mode}.ts")
-}
-
-fn format_filename_timestamp(unix_secs: u64) -> String {
-    let dt = datetime_from_unix(unix_secs);
-
-    let Ok(format) = format_description::parse("[year]-[month]-[day]-[hour][minute]") else {
-        return unix_secs.to_string();
-    };
-
-    dt.format(&format).unwrap_or_else(|_| unix_secs.to_string())
-}
-
-fn sanitize_filename(value: &str) -> String {
-    let mut sanitized = value
-        .chars()
-        .map(|ch| {
-            if ch.is_ascii_alphanumeric() || ch == '-' || ch == '_' {
-                ch
-            } else {
-                '_'
-            }
-        })
-        .collect::<String>()
-        .to_ascii_lowercase();
-
-    while sanitized.contains("__") {
-        sanitized = sanitized.replace("__", "_");
-    }
-
-    sanitized = sanitized.trim_matches('_').to_string();
-    if sanitized.len() > 64 {
-        sanitized.truncate(64);
-    }
-    sanitized
-}
-
-fn list_recording_files(dir: &Path, status: &str, limit: usize) -> Vec<RecordingFileEntry> {
-    let mut entries: Vec<(String, PathBuf)> = Vec::new();
-    collect_recording_files(dir, &mut entries);
-
-    entries.sort_by_key(|(_, path)| {
-        std::cmp::Reverse(
-            fs::metadata(path)
-                .ok()
-                .and_then(|meta| meta.modified().ok())
-                .unwrap_or(SystemTime::UNIX_EPOCH),
-        )
-    });
-
-    entries
-        .into_iter()
-        .take(limit)
-        .map(|(channel_login, path)| RecordingFileEntry {
-            channel_login,
-            filename: path
-                .file_name()
-                .and_then(|f| f.to_str())
-                .unwrap_or("unknown")
-                .to_string(),
-            path_display: path.display().to_string(),
-            status: status.to_string(),
-            pinned: is_recording_pinned(&path),
-        })
-        .collect()
-}
-
-fn collect_recording_files(dir: &Path, out: &mut Vec<(String, PathBuf)>) {
-    let Ok(read) = fs::read_dir(dir) else {
-        return;
-    };
-
-    for entry in read.flatten() {
-        let path = entry.path();
-        if path.is_file() {
-            if !is_visible_recording_file(&path) {
-                continue;
-            }
-            out.push((channel_login_for_recording(&path), path));
-            continue;
-        }
-        if path.is_dir() {
-            collect_recording_files(&path, out);
-        }
-    }
-}
-
-fn channel_login_for_recording(path: &Path) -> String {
-    let parts: Vec<String> = path
-        .components()
-        .map(|component| component.as_os_str().to_string_lossy().to_string())
-        .collect();
-    for (index, part) in parts.iter().enumerate() {
-        if (part == "completed" || part == "incomplete") && index + 1 < parts.len() {
-            return parts[index + 1].clone();
-        }
-    }
-    "unknown".to_string()
-}
-
-fn is_visible_recording_file(path: &Path) -> bool {
-    path.extension()
-        .and_then(|ext| ext.to_str())
-        .map(|ext| {
-            matches!(
-                ext.to_ascii_lowercase().as_str(),
-                "ts" | "mp4" | "mkv" | "m4v" | "mov" | "webm"
-            )
-        })
-        .unwrap_or(false)
-}
-
-fn validate_recording_filename(filename: &str) -> Result<String, RecordingError> {
-    let trimmed = filename.trim();
-    if trimmed.is_empty() {
-        return Err(RecordingError::EmptyFilename);
-    }
-    if trimmed.contains('/') || trimmed.contains('\\') {
-        return Err(RecordingError::InvalidFilename);
-    }
-    if trimmed == "." || trimmed == ".." {
-        return Err(RecordingError::InvalidFilename);
-    }
-
-    Ok(trimmed.to_string())
-}
-
-fn now_unix() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_secs())
-        .unwrap_or(0)
-}
-
-fn move_file_if_exists(from: &Path, to: &Path) -> bool {
-    if !from.exists() {
-        return false;
-    }
-    if let Some(parent) = to.parent() {
-        let _ = fs::create_dir_all(parent);
-    }
-    fs::rename(from, to).is_ok()
-}
-
 fn write_ffmetadata_chapters(path: &Path, events: &[ChapterEvent]) -> Result<(), String> {
     let mut content = String::from(";FFMETADATA1\n");
     for (index, event) in events.iter().enumerate() {
@@ -1368,89 +800,13 @@ fn write_ffmetadata_chapters(path: &Path, events: &[ChapterEvent]) -> Result<(),
     })
 }
 
-fn find_file_by_name_recursive(dir: &Path, filename: &str) -> Option<PathBuf> {
-    let entries = fs::read_dir(dir).ok()?;
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.is_file() {
-            if path.file_name().and_then(|f| f.to_str()) == Some(filename) {
-                return Some(path);
-            }
-            continue;
-        }
-        if path.is_dir()
-            && let Some(found) = find_file_by_name_recursive(&path, filename)
-        {
-            return Some(found);
-        }
-    }
-    None
-}
-
-fn prune_completed_channel_dir(dir: &Path, keep_last: usize) {
-    let mut files: Vec<PathBuf> = Vec::new();
-    collect_recording_media_paths(dir, &mut files);
-
-    files.retain(|path| !is_recording_pinned(path));
-
-    files.sort_by_key(|path| {
-        std::cmp::Reverse(
-            fs::metadata(path)
-                .ok()
-                .and_then(|meta| meta.modified().ok())
-                .unwrap_or(SystemTime::UNIX_EPOCH),
-        )
-    });
-
-    for old_path in files.into_iter().skip(keep_last) {
-        if let Err(error) = fs::remove_file(&old_path) {
-            tracing::warn!(
-                path = %old_path.display(),
-                error = %error,
-                "failed to prune old completed recording"
-            );
-            continue;
-        }
-        let nfo = old_path.with_extension("nfo");
-        if nfo.exists() {
-            let _ = fs::remove_file(nfo);
-        }
-    }
-}
-
-fn pin_marker_path_for_recording(recording_path: &Path) -> PathBuf {
-    let file_name = recording_path
-        .file_name()
-        .and_then(|value| value.to_str())
-        .unwrap_or("recording");
-    recording_path.with_file_name(format!("{file_name}.pin"))
-}
-
-fn is_recording_pinned(recording_path: &Path) -> bool {
-    pin_marker_path_for_recording(recording_path).exists()
-}
-
-fn collect_recording_media_paths(dir: &Path, out: &mut Vec<PathBuf>) {
-    let Ok(entries) = fs::read_dir(dir) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.is_file() {
-            if is_visible_recording_file(&path) {
-                out.push(path);
-            }
-            continue;
-        }
-        if path.is_dir() {
-            collect_recording_media_paths(&path, out);
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use super::RecordingService;
+    use super::files::*;
+    use super::nfo::*;
+    use super::types::*;
+    use std::path::Path;
 
     #[test]
     fn parse_display_episode_suffix_handles_base_and_indexed() {
